Merge Details:

Vincent Thiberville (1):
      97171072ad cflags: get proper gcc version with gcc >= 7

Xavier Chantry (1):
      a0d64eb8c4 support gcc 8

Conflicts:
 Config/cflags.sh

+++ str-l.h
@@ -409,7 +409,11 @@ static inline lstr_t r_lstr_cat(const lstr_t s1, const lstr_t s2)
 /** Returns "memcmp" ordering of \v s1 and \v s2. */
 static ALWAYS_INLINE int lstr_cmp(const lstr_t s1, const lstr_t s2)
 {
-    int len = MIN(s1.len, s2.len);
+    /* workaround for a warning of -Wstringop-overflow in gcc 8
+     * see https://gcc.gnu.org/bugzilla/show_bug.cgi?id=89699
+     */
+    size_t len = MIN((size_t)s1.len, (size_t)s2.len);
+
     return memcmp(s1.s, s2.s, len) ?: CMP(s1.len, s2.len);
 }
